HAVE_WebApp/requirements.txt
appdirs==1.4.0
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
mysqlclient==2.0.3
packaging==16.8
Pillow==8.1.2
pyparsing==2.1.10
pytz==2021.1
six==1.10.0
sqlparse==0.4.1